+++ .travis.yml
@@ -7,7 +7,7 @@ services:
 script:
 - pip  install --upgrade setuptools wheel boto3 aioboto3 moto pytest requests pytest-asyncio
 - pip install -q localstack; docker pull localstack/localstack; localstack start --docker
-  & sleep 20
+  & sleep 30
 - export AWS_ACCESS_KEY_ID=foobar AWS_SECRET_ACCESS_KEY=foobar AWS_DEFAULT_REGION=us-east-1
 - pytest tests
 deploy:
